refactor(cli): extract manifest channel domain
